Move tilequeue.postgresql -> tilequeue.query.pool
The db connection pool logic seemed like a good candidate for moving
under the tilequeue.query location.
